// src/app/types.ts
import {requestStatus} from '../enum/requestStatus';

export type RequestStatusType = requestStatus;
export type InitialStateType = {
  // происходит ли сейчас взаимодействие с сервером
  status: RequestStatusType
  // если ошибка какая-то глобальная произойдёт - мы запишем текст ошибки сюда
  error: string | null
  isInitialized: boolean
}